requirements.txt
requests>=2.25.1
tqdm>=4.60.0
python-dotenv
